src/scenario-05-batch-processing/requirements.txt
vibevoice[streamingtts] @ git+https://github.com/microsoft/VibeVoice.git
click
tqdm
pyyaml
soundfile
